fix nullable C# 8.0 test be ignored on unless custom flag is defined
Standard project settings use compiler C# 7.3. Testing
a new feature of C# 8.0 is not possible with this settings.
Hence, testing the new nullable reference feature is
not possible.

In order to test this, you need to enable C# 8.0 and define a custom
constant in the project settings. The constant is named
`CUSTOM_FLAG_NULLABLE_AVAILABLE`