refact:(ecxpetion) pass inner exception to error
It is valuable to tell, where the exception occurred.
So, whenever an exception is cached, it should be
passed to the newly created exception as
"inner exception".

This is done now at some places, where the custom
exception `ErrorMessage` is created. It will help a great
deal to get to the cause of the error.